@@ -2,8 +2,8 @@ package com.edricchan.androidexamples.ui.components;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
 import com.edricchan.androidexamples.R;
 
 public class ComponentsRaisedButtonActivity extends AppCompatActivity {
