@@ -1,10 +1,10 @@
 package com.edricchan.androidexamples.ui.home;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.design.widget.BottomNavigationView;
-import android.support.v4.app.Fragment;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.NonNull;
+import com.google.android.material.bottomnavigation.BottomNavigationView;
+import androidx.fragment.app.Fragment;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.MenuItem;
 import com.edricchan.androidexamples.R;
 import com.edricchan.androidexamples.ui.fragments.home.ComponentsFragment;
